noop separate upgrade.go, upgrade_rp.go
